% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{chipseq-cbp}
\alias{chipseq-cbp}
\alias{cbp-csaw-benchmark}
\alias{cbp-csaw-uninf-benchmark}
\alias{cbp-csaw-cov-benchmark}
\title{ChIP-seq, CBP dataset results}
\description{
Results for case study on differential peak calling in CREB-binding protein (CBP)
knockout and wild-type mice described in Korthauer and Kimes et al. (2019).
Analyses were performed using the \code{csaw} package.
}
\details{
Results for this case study include:
\itemize{
\item \code{cbp-csaw-benchmark}: using csaw with region width as covariate.
\item \code{cbp-csaw-uninf-benchmark}: using csaw with uninformative covariate.
\item \code{cbp-csaw-cov-benchmark}: using csaw with mean coverage as covariate.
}
}
\examples{
\dontrun{
`cbp-csaw-benchmark`()
`cbp-csaw-uninf-benchmark`()
`cbp-csaw-cov-benchmark`()
}

}
\seealso{
\code{\link{chipseq-h3k4me3}}
}
