% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{chipseq-h3k4me3}
\alias{chipseq-h3k4me3}
\alias{h3k4me3-promoters-benchmark}
\alias{h3k4me3-csaw-benchmark}
\alias{h3k4me3-csaw-uninf-benchmark}
\alias{h3k4me3-csaw-cov-benchmark}
\title{ChIP-seq, H3K4me3 dataset results}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for case study on differential peak calling in human H3K4me3
ChIP-seq data from ENCODE described in Korthauer and Kimes et al. (2019).
Analyses were performed using either \code{DESeq2}
applied to predefined regions or with the \code{csaw} package.
}
\details{
Results for this case study include:
\itemize{
\item \code{h3k4me3-promoters-benchmark}: using predefined regions with mean coverage as covariate.
\item \code{h3k4me3-csaw-benchmark}: using csaw with region width as covariate.
\item \code{h3k4me3-csaw-uninf-benchmark}: using csaw with uninformative covariate.
\item \code{h3k4me3-csaw-cov-benchmark}: using csaw with mean coverage as covariate.
}
}
\examples{
`h3k4me3-promoters-benchmark`(metadata = TRUE)

\dontrun{
`h3k4me3-promoters-benchmark`()
`h3k4me3-csaw-benchmark`()
`h3k4me3-csaw-uninf-benchmark`()
`h3k4me3-csaw-cov-benchmark`()
}

}
\seealso{
\code{\link{chipseq-cbp}}
}
