% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{gsa-human}
\alias{gsa-human}
\alias{human-benchmark}
\alias{human-uninf-benchmark}
\title{Gene Set Analysis, human dataset results}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for case study on gene set analysis using human dataset described in
Korthauer and Kimes et al. (2019). Analyses were performed using GSEA.
}
\details{
Results for this case study include:
\itemize{
\item \code{human-benchmark}: using gene set size as covariate.
\item \code{human-uninf-benchmark}: using uninformative covariate.
}
}
\examples{
`human-benchmark`(metadata = TRUE)

\dontrun{
`human-benchmark`()
`human-uninf-benchmark`()
}

}
\seealso{
\code{\link{gsa-human}}
}
