% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{gsa-mouse}
\alias{gsa-mouse}
\alias{mouse-benchmark}
\alias{mouse-uninf-benchmark}
\title{Gene Set Analysis, mouse dataset results}
\description{
Results for case study on gene set analysis using mouse dataset described in
Korthauer and Kimes et al. (2019). Analyses were performed using GSEA.
}
\details{
Results for this case study include:
\itemize{
\item \code{mouse-benchmark}: using gene set size as covariate.
\item \code{mouse-uninf-benchmark}: using uninformative covariate.
}
}
\examples{
res <- `mouse-benchmark`()

\dontrun{
`mouse-benchmark`()
`mouse-uninf-benchmark`()
}

}
\seealso{
\code{\link{gsa-human}}
}
