% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{gwas-bmi}
\alias{gwas-bmi}
\alias{bmi-maf-benchmark}
\alias{bmi-uninf-benchmark}
\alias{bmi-samplesize-benchmark}
\title{Genome-wide Association Study, BMI dataset results}
\description{
Results for case study on genome-wide associate study meta-analysis using a BMI dataset
from the GIANT Consortium 
described in Korthauer and Kimes et al. (2019).
}
\details{
Results for this case study include:
\itemize{
\item \code{bmi-maf-benchmark}: using minor allele frequency as covariate.
\item \code{bmi-uninf-benchmark}: using uninformative covariate.
\item \code{bmi-samplesize-benchmark}: using sample size as covariate.
}
}
\examples{
res <- `bmi-maf-benchmark`()

\dontrun{
`bmi-maf-benchmark`()
`bmi-uninf-benchmark`()
`bmi-samplesize-benchmark`()
}

}
