% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{microbiome-enigma}
\alias{microbiome-enigma}
\alias{enigma-al-otus-benchmark}
\alias{enigma-al-otus-abun-benchmark}
\alias{enigma-al-otus-uninf-benchmark}
\alias{enigma-ph-otus-benchmark}
\alias{enigma-ph-otus-abun-benchmark}
\alias{enigma-ph-otus-uninf-benchmark}
\alias{enigma-so4-otus-benchmark}
\alias{enigma-so4-otus-abun-benchmark}
\alias{enigma-so4-otus-uninf-benchmark}
\title{Microbiome, ENIGMA dataset results}
\description{
Results for case study on 16S microbiome Spearman correlation testing of
ENIGMA dataset described in Korthauer and
Kimes et al. (2019). Spearman correlation test was used to test
for correlation between OTU-level microbial abundances and various geochemical
and physical measurements.
}
\details{
Results for this case study include:
\itemize{
\item \code{enigma-al-otus-benchmark}: testing for correlation with Al measurements using ubiquity as covariate.
\item \code{enigma-al-otus-abun-benchmark}: testing for correlation with Al measurements using mean abundance as covariate.
\item \code{enigma-al-otus-uninf-benchmark}: testing for correlation with Al measurements using uninformative covariate.
\item \code{enigma-ph-otus-benchmark}: testing for correlation with pH measurements using ubiquity as covariate.
\item \code{enigma-ph-otus-abun-benchmark}: testing for correlation with pH measurements using mean abundance as covariate.
\item \code{enigma-ph-otus-uninf-benchmark}: testing for correlation with pH measurements using uninformative covariate.
\item \code{enigma-so4-otus-benchmark}: testing for correlation with SO4 measurements using ubiquity as covariate.
\item \code{enigma-so4-otus-abun-benchmark}: testing for correlation with SO4 measurements using mean abundance as covariate.
\item \code{enigma-so4-otus-uninf-benchmark}: testing for correlation with SO4 measurements using uninformative covariate.
}
}
\examples{
\dontrun{
`enigma-al-otus-benchmark`()
`enigma-al-otus-abun-benchmark`()
`enigma-al-otus-uninf-benchmark`()
`enigma-ph-otus-benchmark`()
`enigma-ph-otus-abun-benchmark`()
`enigma-ph-otus-uninf-benchmark`()
`enigma-so4-otus-benchmark`()
`enigma-so4-otus-abun-benchmark`()
`enigma-so4-otus-uninf-benchmark`()
}

}
\seealso{
\code{\link{microbiome-schubert}}, \code{\link{microbiome-baxter}}, \code{\link{microbiome-goodrich}}, \code{\link{microbiome-papa}}
}
