% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{microbiome-goodrich}
\alias{microbiome-goodrich}
\alias{goodrich-genus-benchmark}
\alias{goodrich-genus-abun-benchmark}
\alias{goodrich-genus-uninf-benchmark}
\alias{goodrich-otus-benchmark}
\title{Microbiome, Goodrich dataset results}
\description{
Results for case study on 16S microbiome differential testing of
Goodrich et al. (2014) dataset described in Korthauer and
Kimes et al. (2019). Wilcoxon rank-sum test was used to test
for differential abundance between samples from lean (body mass index, BMI < 25)
and obsese (BMI > 30) individuals.
}
\details{
Results for this case study include:
\itemize{
\item \code{goodrich-genus-benchmark}: testing at genus-level using ubiquity as covariate.
\item \code{goodrich-genus-abun-benchmark}: testing at genus-level using mean abundance as covariate.
\item \code{goodrich-genus-uninf-benchmark}: testing at genus-level using uninformative covariate.
\item \code{goodrich-otus-benchmark}: testing at OTU-level using ubiquity as covariate.
}
}
\examples{
`goodrich-genus-benchmark`(metadata = TRUE)

\dontrun{
`goodrich-genus-benchmark`()
`goodrich-genus-abun-benchmark`()
`goodrich-genus-uninf-benchmark`()
`goodrich-otus-benchmark`()
}

}
\references{
\itemize{
\item Goodrich, J. K. et al. (2014). Human genetics shape the gut microbiome. Cell, 159(4), 789-799.
}
}
\seealso{
\code{\link{microbiome-schubert}}, \code{\link{microbiome-baxter}}, \code{\link{microbiome-papa}}, \code{\link{microbiome-enigma}}
}
