% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{microbiome-papa}
\alias{microbiome-papa}
\alias{papa-genus-benchmark}
\alias{papa-genus-abun-benchmark}
\alias{papa-genus-log-ubiquity-benchmark}
\alias{papa-genus-uninf-benchmark}
\alias{papa-otus-benchmark}
\title{Microbiome, Papa dataset results}
\description{
Results for case study on 16S microbiome differential testing of
Papa et al. (2012) dataset described in Korthauer and
Kimes et al. (2019). Wilcoxon rank-sum test was used to test
for differential abundance between samples from patients
with GI symptoms but not IBD, and IBD patients.
}
\details{
Results for this case study include:
\itemize{
\item \code{papa-genus-benchmark}: testing at genus-level using ubiquity as covariate.
\item \code{papa-genus-abun-benchmark}: testing at genus-level using mean abundance as covariate.
\item \code{papa-genus-log-ubiquity-benchmark}: testing at genus-level using log-ubiquity as covariate.
\item \code{papa-genus-uninf-benchmark}: testing at genus-level using uninformative covariate.
\item \code{papa-otus-benchmark}: testing at OTU-level using ubiquity as covariate.
}
}
\examples{
\dontrun{
`papa-genus-benchmark`()
`papa-genus-abun-benchmark`()
`papa-genus-log-ubiquity-benchmark`()
`papa-genus-uninf-benchmark`()
`papa-otus-benchmark`()
}

}
\references{
\itemize{
\item Papa, E. et al. (2012). Non-invasive mapping of the gastrointestinal microbiota identifies children with inflammatory bowel disease. PloS One, 7(6), e39242.
}
}
\seealso{
\code{\link{microbiome-schubert}}, \code{\link{microbiome-baxter}}, \code{\link{microbiome-goodrich}}, \code{\link{microbiome-enigma}}
}
