% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{microbiome-schubert}
\alias{microbiome-schubert}
\alias{schubert-otus-benchmark}
\alias{schubert-otus-abun-benchmark}
\alias{schubert-otus-uninf-benchmark}
\title{Microbiome, Schubert dataset results}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for case study on 16S microbiome differential testing of
Schubert et al. (2014) dataset described in Korthauer and
Kimes et al. (2019). Wilcoxon rank-sum test was used to test
for differential abundance between samples from healthy patients
and patients with diarrheal stool samples (including both C. difficile infection,
CDI, and non-CDI samples).
}
\details{
Results for this case study include:
\itemize{
\item \code{schubert-otus-benchmark}: testing at OTU-level using ubiquity as covariate.
\item \code{schubert-otus-abun-benchmark}: testing at OTU-level using mean abundance as covariate.
\item \code{schubert-otus-uninf-benchmark}: testing at OTU-level using uninformative covariate.
}
}
\examples{
`schubert-otus-benchmark`(metadata = TRUE)

\dontrun{
`schubert-otus-benchmark`()
`schubert-otus-abun-benchmark`()
`schubert-otus-uninf-benchmark`()
}

}
\references{
\itemize{
\item Schubert, A. M. et al. (2014). Microbiome data distinguish patients with Clostridium difficile infection and non-C. difficile-associated diarrhea from healthy controls. MBio, 5(3), e01021-14.
}
}
\seealso{
\code{\link{microbiome-baxter}}, \code{\link{microbiome-goodrich}}, \code{\link{microbiome-papa}}, \code{\link{microbiome-enigma}}
}
