% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{polyester-insilico}
\alias{polyester-insilico}
\alias{polyester-results-de5}
\alias{polyester-results-de5-uninfCov}
\alias{polyester-results-de10}
\alias{polyester-results-de10-uninfCov}
\alias{polyester-results-null5}
\alias{polyester-results-null10}
\alias{polyesterW-results-de5}
\alias{polyesterW-results-de10}
\title{Polyester simulated dataset results}
\description{
Results for differential gene expression analysis using 
RNA-seq data simulated with the \pkg{Polyester} package
from yeast data obtained from Schurch et al. (2016) as 
described in Korthauer and Kimes et al. (2019).
Differential expression was tested using \pkg{DESeq2}.
}
\details{
Results for this case study include:
\itemize{
\item \code{polyester-results-null5}: null 5 vs. 5 comparison.
\item \code{polyester-results-de5}: using unimodal alternative, 2000 DE genes, strong informative covariate, 5 vs. 5 comparison.
\item \code{polyester-results-de5-uninfCov}: using unimodal alternative, 2000 DE genes, uninformative covariate, 5 vs. 5 comparison.
\item \code{polyesterW-results-de5}: using unimodal alternative, 2000 DE genes, weak informative covariate, 5 vs. 5 comparison.
\item \code{polyester-results-null10}: null 10 vs. 10 comparison.
\item \code{polyester-results-de10}: using unimodal alternative, 2000 DE genes, strong informative covariate, 10 vs. 10 comparison.
\item \code{polyester-results-de10-uninfCov}: using unimodal alternative, 2000 DE genes, uninformative covariate, 10 vs. 10 comparison.
\item \code{polyesterW-results-de10}: using unimodal alternative, 2000 DE genes, weak informative covariate, 10 vs. 10 comparison.
}
}
\examples{
`polyester-results-de5`(metadata = TRUE)

\dontrun{
`polyester-results-de5`()
`polyester-results-de5-uninfCov`()
`polyester-results-de10`()
`polyester-results-de10-uninfCov`()
`polyester-results-null5`()
`polyester-results-null10`()
`polyesterW-results-de5`()
`polyesterW-results-de10`()
}

}
\references{
\itemize{
\item Schurch, N. J. et al. (2016). How many biological replicates are needed in an RNA-seq experiment and which differential expression tool should you use?. RNA, 22(6), 839-851.
}
}
\seealso{
\code{\link{yeast-insilico}}
}
