% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{rnaseq-brain}
\alias{rnaseq-brain}
\alias{brain-benchmark}
\alias{brain-uninf-benchmark}
\title{RNA-seq, brain dataset results}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for case study on differential gene expression analysis between
two brain regions (nucleus accumbens and putamen) using data from
the GTEx project described in Korthauer and Kimes et al. (2019).
Differential expression was tested using \pkg{DESeq2}.
}
\details{
Results for this case study include:
\itemize{
\item \code{brain-benchmark}: using mean counts as covariate.
\item \code{brain-uninf-benchmark}: using uninformative covariate.
}
}
\examples{
`brain-benchmark`(metadata = TRUE)

\dontrun{
`brain-benchmark`()
`brain-uninf-benchmark`()
}

}
\seealso{
\code{\link{rnaseq-mir200c}}
}
