% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{rnaseq-mir200c}
\alias{rnaseq-mir200c}
\alias{mir200c-benchmark}
\alias{mir200c-uninf-benchmark}
\title{RNA-seq, mir200c dataset results}
\description{
Results for case study on differential gene expression analysis between
mir200c knockdown cells and untreated cells described in
Korthauer and Kimes et al. (2019).
Differential expression was tested using \pkg{DESeq2}.
}
\details{
Results for this case study include:
\itemize{
\item \code{mir200c-benchmark}: using mean counts as covariate.
\item \code{mir200c-uninf-benchmark}: using uninformative covariate.
}
}
\examples{
`mir200c-benchmark`(metadata = TRUE)

\dontrun{
`mir200c-benchmark`()
`mir200c-uninf-benchmark`()
}

}
\seealso{
\code{\link{rnaseq-brain}}
}
