% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{scrnaseq-human}
\alias{scrnaseq-human}
\alias{human-benchmark-mast-det}
\alias{human-benchmark-mast-mean}
\alias{human-benchmark-mast-uninf}
\alias{human-benchmark-scdd-det}
\alias{human-benchmark-scdd-mean}
\alias{human-benchmark-scdd-uninf}
\alias{human-benchmark-wilcox-det}
\alias{human-benchmark-wilcox-mean}
\alias{human-benchmark-wilcox-uninf}
\title{scRNA-seq, human dataset results}
\description{
Results for case study on single-cell RNA-seq differential expression analysis
using data from diffuse neoplastic infiltrating cells at the migrating front
of human glioblastoma as described in Korthauer and Kimes et al. (2019).
Differential expression was testing using either the \pkg{MAST} package,
\pkg{scDD} package, or Wilcoxon rank-sum test.
}
\details{
Results for this case study include:
\itemize{
\item \code{human-benchmark-mast-det}: testing with \pkg{MAST}, using detection rate as covariate.
\item \code{human-benchmark-mast-mean}: testing with \pkg{MAST}, using mean counts as covariate.
\item \code{human-benchmark-mast-uninf}: testing with \pkg{MAST}, using uninformative covariate.
\item \code{human-benchmark-scdd-det}: testing with \pkg{scDD}, using detection rate as covariate.
\item \code{human-benchmark-scdd-mean}: testing with \pkg{scDD}, using mean counts as covariate.
\item \code{human-benchmark-scdd-uninf}: testing with \pkg{scDD}, using uninformative covariate.
\item \code{human-benchmark-wilcox-det}: testing with Wilcoxon rank-sum, using detection rate as covariate.
\item \code{human-benchmark-wilcox-mean}: testing with Wilcoxon rank-sum, using mean counts as covariate.
\item \code{human-benchmark-wilcox-uninf}: testing with Wilcoxon rank-sum, using uninformative covariate.
}
}
\examples{
res <- `human-benchmark-mast-det`()

\dontrun{
`human-benchmark-mast-det`()
`human-benchmark-mast-mean`()
`human-benchmark-mast-uninf`()
`human-benchmark-scdd-det`()
`human-benchmark-scdd-mean`()
`human-benchmark-scdd-uninf`()
`human-benchmark-wilcox-det`()
`human-benchmark-wilcox-mean`()
`human-benchmark-wilcox-uninf`()
}

}
\seealso{
\code{\link{scrnaseq-mouse}}
}
