% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{scrnaseq-mouse}
\alias{scrnaseq-mouse}
\alias{mouse-benchmark-mast-det}
\alias{mouse-benchmark-mast-mean}
\alias{mouse-benchmark-mast-uninf}
\alias{mouse-benchmark-scdd-det}
\alias{mouse-benchmark-scdd-mean}
\alias{mouse-benchmark-scdd-uninf}
\alias{mouse-benchmark-wilcox-det}
\alias{mouse-benchmark-wilcox-mean}
\alias{mouse-benchmark-wilcox-uninf}
\title{scRNA-seq, mouse dataset results}
\description{
Results for case study on single-cell RNA-seq differential expression analysis
using data from mouse cells at NF-kappa-beta activation as described in
Korthauer and Kimes et al. (2019).
Differential expression was testing using either the \pkg{MAST} package,
\pkg{scDD} package, or Wilcoxon rank-sum test.
}
\details{
Results for this case study include:
\itemize{
\item \code{mouse-benchmark-mast-det}: testing with \pkg{MAST}, using detection rate as covariate.
\item \code{mouse-benchmark-mast-mean}: testing with \pkg{MAST}, using mean counts as covariate.
\item \code{mouse-benchmark-mast-uninf}: testing with \pkg{MAST}, using uninformative covariate.
\item \code{mouse-benchmark-scdd-det}: testing with \pkg{scDD}, using detection rate as covariate.
\item \code{mouse-benchmark-scdd-mean}: testing with \pkg{scDD}, using mean counts as covariate.
\item \code{mouse-benchmark-scdd-uninf}: testing with \pkg{scDD}, using uninformative covariate.
\item \code{mouse-benchmark-wilcox-det}: testing with Wilcoxon rank-sum, using detection rate as covariate.
\item \code{mouse-benchmark-wilcox-mean}: testing with Wilcoxon rank-sum, using mean counts as covariate.
\item \code{mouse-benchmark-wilcox-uninf}: testing with Wilcoxon rank-sum, using uninformative covariate.
}
}
\examples{
`mouse-benchmark-mast-det`(metadata = TRUE)

\dontrun{
`mouse-benchmark-mast-det`()
`mouse-benchmark-mast-mean`()
`mouse-benchmark-mast-uninf`()
`mouse-benchmark-scdd-det`()
`mouse-benchmark-scdd-mean`()
`mouse-benchmark-scdd-uninf`()
`mouse-benchmark-wilcox-det`()
`mouse-benchmark-wilcox-mean`()
`mouse-benchmark-wilcox-uninf`()
}

}
\seealso{
\code{\link{scrnaseq-human}}
}
