% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{sims-informative-cubic}
\alias{sims-informative-cubic}
\alias{informative-cubic-benchmark-chisq4}
\alias{informative-cubic-benchmark-gaussian}
\alias{informative-cubic-benchmark-t11}
\alias{informative-cubic-benchmark-t5}
\title{Simulation study results (cubic covariate)}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for simulations used in Korthauer and Kimes et al. (2019) with `cubic'
informative and paired uninformative covariates stored
as a list of SummarizedBenchmark objects from 100 replications. 
Simulations were performed with standard Gaussian, \eqn{t_5}, \eqn{t_{11}},
and \eqn{\chi_4} distributed test statistics.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
`informative-cubic-benchmark-gaussian`(metadata = TRUE)

\dontrun{
`informative-cubic-benchmark-gaussian`()
`informative-cubic-benchmark-t11`()
`informative-cubic-benchmark-t5`()
`informative-cubic-benchmark-chisq4`()

res <- `informative-cubic-benchmark-chisq4`()
res[[1]]$informative
res[[1]]$uninformative
}

}
