% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{sims-null}
\alias{sims-null}
\alias{null-benchmark-chisq4}
\alias{null-benchmark-gaussian}
\alias{null-benchmark-t11}
\alias{null-benchmark-t5}
\title{Simulation study results (null settings)}
\arguments{
\item{metadata}{logical whether to load metadata (TRUE) or full resource (FALSE).
(default = FALSE)}
}
\description{
Results for null simulations used in Korthauer and Kimes et al. (2019) stored
as a list of SummarizedBenchmark objects from 100 replications. 
Simulations were performed with standard Gaussian, \eqn{t_5}, \eqn{t_{11}},
and \eqn{\chi_4} distributed test statistics.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
`null-benchmark-gaussian`(metadata = TRUE)

\dontrun{
`null-benchmark-gaussian`()
`null-benchmark-t11`()
`null-benchmark-t5`()
`null-benchmark-chisq4`()

res <- `null-benchmark-chisq4`()
res[[1]]$informative
res[[1]]$uninformative
}

}
