% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{sims-varyinginfo-discrete}
\alias{sims-varyinginfo-discrete}
\alias{varyinginfo-discrete-benchmark-level00}
\alias{varyinginfo-discrete-benchmark-level20}
\alias{varyinginfo-discrete-benchmark-level40}
\alias{varyinginfo-discrete-benchmark-level60}
\alias{varyinginfo-discrete-benchmark-level80}
\alias{varyinginfo-discrete-benchmark-level100}
\title{Simulation study results (varying information, discrete)}
\description{
Results for simulations used in Korthauer and Kimes et al. (2019) across varying
covariate informativeness, using standard Gaussian distributed test statistics,
and informative and paired uninformative covariates stored
as a list of SummarizedBenchmark objects from 100 replications. 
An informative covariate with tunable informativeness was defined with a
rough informativeness parameter, delta, defined between 0 and 100.
Simulations were performed with delta values of 0, 20, 40, 60, 80, and 100.
The conditional probability of a test being null was
defined by a discrete, non-smooth, function of the informative covariate.

In \code{\link{sims-varyinginfo-smooth}},
similar settings were considered, but using a smooth relationship between the
condition probability of a test being null and the informative covariate.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
\dontrun{
`varyinginfo-discrete-benchmark-level00`()
`varyinginfo-discrete-benchmark-level20`()
`varyinginfo-discrete-benchmark-level40`()
`varyinginfo-discrete-benchmark-level60`()
`varyinginfo-discrete-benchmark-level80`()
`varyinginfo-discrete-benchmark-level100`()

res <- `varyinginfo-discrete-benchmark-level100`()
res[[1]]$informative
res[[1]]$uninformative
}

}
\seealso{
\code{\link{sims-varyinginfo-smooth}}
}
