% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{sims-varyingntests}
\alias{sims-varyingntests}
\alias{varyingntests-benchmark-n100}
\alias{varyingntests-benchmark-n500}
\alias{varyingntests-benchmark-n1000}
\alias{varyingntests-benchmark-n5000}
\alias{varyingntests-benchmark-n10000}
\alias{varyingntests-benchmark-n50000}
\title{Simulation study results (varying number of tests)}
\description{
Results for simulations used in Korthauer and Kimes et al. (2019) across varying
number of hypotheses tested, using standard Gaussian distributed test statistics,
and `sine' informative and paired uninformative covariates stored
as a list of SummarizedBenchmark objects from 100 replications. 
Simulations were performed with 100, 500, 1000, 5000, 10000, and 50000 tests.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
`varyingntests-benchmark-n100`(metadata = TRUE)

\dontrun{
`varyingntests-benchmark-n100`()
`varyingntests-benchmark-n500`()
`varyingntests-benchmark-n1000`()
`varyingntests-benchmark-n5000`()
`varyingntests-benchmark-n10000`()
`varyingntests-benchmark-n50000`()

res <- `varyingntests-benchmark-n100`()
res[[1]]$informative
res[[1]]$uninformative
}

}
