% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{sims-varyingpi0}
\alias{sims-varyingpi0}
\alias{varyingpi0-benchmark-nullprop05}
\alias{varyingpi0-benchmark-nullprop10}
\alias{varyingpi0-benchmark-nullprop20}
\alias{varyingpi0-benchmark-nullprop30}
\alias{varyingpi0-benchmark-nullprop40}
\alias{varyingpi0-benchmark-nullprop50}
\alias{varyingpi0-benchmark-nullprop60}
\alias{varyingpi0-benchmark-nullprop70}
\alias{varyingpi0-benchmark-nullprop80}
\alias{varyingpi0-benchmark-nullprop90}
\alias{varyingpi0-benchmark-nullprop95}
\alias{varyingpi0-benchmark-nullprop99}
\title{Simulation study results (varying null proportion, Gaussian-distributed)}
\description{
Results for simulations used in Korthauer and Kimes et al. (2019) across varying
null proportion of hypotheses with 20000 tests, standard Gaussian distributed
test statistics, and `sine' informative and paired uninformative covariates stored
as a list of SummarizedBenchmark objects from 100 replications. 
Simulations were performed with 5\%, 10\%, 20\%, 30\%, 40\%, 50\%, 60\%, 70\%,
80\%, 90\%, 95\%, and 99\% null tests.

In \code{\link{sims-varyingpi0-t}},
similar settings were considered, but \eqn{t_{11}} distributed test statistics.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
res <- `varyingpi0-benchmark-nullprop99`()
res[[1]]$informative
res[[1]]$uninformative

\dontrun{
`varyingpi0-benchmark-nullprop05`()
`varyingpi0-benchmark-nullprop10`()
`varyingpi0-benchmark-nullprop20`()
`varyingpi0-benchmark-nullprop30`()
`varyingpi0-benchmark-nullprop40`()
`varyingpi0-benchmark-nullprop50`()
`varyingpi0-benchmark-nullprop60`()
`varyingpi0-benchmark-nullprop70`()
`varyingpi0-benchmark-nullprop80`()
`varyingpi0-benchmark-nullprop90`()
`varyingpi0-benchmark-nullprop95`()
`varyingpi0-benchmark-nullprop99`()
}

}
\seealso{
\code{\link{sims-varyingpi0-t}}
}
