% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-simulations.R
\docType{data}
\name{uasettings-benchmark}
\alias{uasettings-benchmark}
\alias{uasettings-benchmark-bimodal}
\alias{uasettings-benchmark-flattop}
\alias{uasettings-benchmark-skew}
\alias{uasettings-benchmark-spiky}
\title{Simulation study results (unimodal, Gaussian-distributed)}
\description{
Results for simulations used in Korthauer and Kimes et al. (2019)
with 'unimodal' series of effect size distributions (Stephens, 2017)
with standard Gaussian distributed test statistics, and `cubic' informative
and paired uninformative covariates stored
as a list of SummarizedBenchmark objects from 100 replications.
In each simulation, 10\% of tests were non-null, i.e. had non-zero effect sizes.
Simulations were performed with `bimodal', `flat top', `skew' and `spiky'
effect size distributions (see corresponding reference for details).

In \code{\link{uasettings-t-benchmark}} and \code{\link{uasettings-25-benchmark}},
similar settings were considered, but with \eqn{t_{11}} distributed test statistics
and 25\% of tests being non-null, respectively.
}
\details{
The object is a list of length 100 corresponding to the 100 replications. Each
entry in the list is a named list of two SummarizedBenchmark objects. The results
based on using an informative and uninformative covariate are stored as
\code{"informative"} and \code{"uninformative"}, respectively.
}
\examples{
\dontrun{
`uasettings-benchmark-bimodal`()
`uasettings-benchmark-flattop`()
`uasettings-benchmark-skew`()
`uasettings-benchmark-spiky`()

res <- `uasettings-benchmark-spiky`()
res[[1]]$informative
res[[1]]$uninformative
}

}
\references{
\itemize{
\item Stephens, M. (2017). False discovery rates: a new deal. Biostatistics, 18(2), 275-294.
}
}
\seealso{
\code{\link{uasettings-t-benchmark}}, \code{\link{uasettings-25-benchmark}}
}
