% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-casestudies.R
\docType{data}
\name{yeast-insilico}
\alias{yeast-insilico}
\alias{yeast-results-de5}
\alias{yeast-results-de5-uninfCov}
\alias{yeast-results-de10}
\alias{yeast-results-de10-uninfCov}
\alias{yeast-results-null5}
\alias{yeast-results-null10}
\alias{yeastH-results-de5}
\alias{yeastH-results-de5-uninfCov}
\alias{yeastH-results-de10}
\alias{yeastH-results-de10-uninfCov}
\alias{yeastHW-results-de5}
\alias{yeastHW-results-de10}
\alias{yeastII-results-de5}
\alias{yeastII-results-de5-uninfCov}
\alias{yeastII-results-de10}
\alias{yeastII-results-de10-uninfCov}
\alias{yeastIIH-results-de5}
\alias{yeastIIH-results-de5-uninfCov}
\alias{yeastIIH-results-de10}
\alias{yeastIIH-results-de10-uninfCov}
\alias{yeastIIHW-results-de5}
\alias{yeastIIHW-results-de10}
\alias{yeastIIW-results-de5}
\alias{yeastIIW-results-de10}
\alias{yeastW-results-de5}
\alias{yeastW-results-de10}
\title{Yeast in silico dataset results}
\description{
Results for differential gene expression analysis using 
in silico (computationally) spiked-in data generated from   
yeast data obtained from Schurch et al. (2016) as 
described in Korthauer and Kimes et al. (2019).
Differential expression was tested using \pkg{DESeq2}.
}
\details{
Results for this case study include:
\itemize{
\item \code{yeast-results-null5}: null 5 vs. 5 comparison.
\item \code{yeast-results-de5}: using unimodal alternative, 2000 DE genes, strong informative covariate, 5 vs. 5 comparison.
\item \code{yeastH-results-de5}: using unimodal alternative, 500 DE genes, strong informative covariate, 5 vs. 5 comparison.
\item \code{yeastII-results-de5}: using bimodal alternative, 2000 DE genes, strong informative covariate, 5 vs. 5 comparison.
\item \code{yeastIIH-results-de5}: using bimodal alternative, 500 DE genes, strong informative covariate, 5 vs. 5 comparison.
\item \code{yeastW-results-de5}: using unimodal alternative, 2000 DE genes, weak informative covariate, 5 vs. 5 comparison.
\item \code{yeastHW-results-de5}: using unimodal alternative, 500 DE genes, weak informative covariate, 5 vs. 5 comparison.
\item \code{yeastIIW-results-de5}: using bimodal alternative, 2000 DE genes, weak informative covariate, 5 vs. 5 comparison.
\item \code{yeastIIHW-results-de5}: using bimodal alternative, 500 DE genes, weak informative covariate, 5 vs. 5 comparison.
\item \code{yeast-results-de5-uninfCov}: using unimodal alternative, 2000 DE genes, uninformative covariate, 5 vs. 5 comparison.
\item \code{yeastH-results-de5-uninfCov}: using unimodal alternative, 500 DE genes, uninformative covariate, 5 vs. 5 comparison.
\item \code{yeastII-results-de5-uninfCov}: using bimodal alternative, 2000 DE genes, uninformative covariate, 5 vs. 5 comparison.
\item \code{yeastIIH-results-de5-uninfCov}: using bimodal alternative, 500 DE genes, uninformative covariate, 5 vs. 5 comparison.
\item \code{yeast-results-null10}: null 10 vs. 10 comparison.
\item \code{yeast-results-de10}: using unimodal alternative, 2000 DE genes, strong informative covariate, 10 vs. 10 comparison.
\item \code{yeastH-results-de10}: using unimodal alternative, 500 DE genes, strong informative covariate, 10 vs. 10 comparison.
\item \code{yeastII-results-de10}: using bimodal alternative, 2000 DE genes, strong informative covariate, 10 vs. 10 comparison.
\item \code{yeastIIH-results-de10}: using bimodal alternative, 500 DE genes, strong informative covariate, 10 vs. 10 comparison.
\item \code{yeastW-results-de10}: using unimodal alternative, 2000 DE genes, weak informative covariate, 10 vs. 10 comparison.
\item \code{yeastHW-results-de10}: using unimodal alternative, 500 DE genes, weak informative covariate, 10 vs. 10 comparison.
\item \code{yeastIIW-results-de10}: using bimodal alternative, 2000 DE genes, weak informative covariate, 10 vs. 10 comparison.
\item \code{yeastIIHW-results-de10}: using bimodal alternative, 500 DE genes, weak informative covariate, 10 vs. 10 comparison.
\item \code{yeast-results-de10-uninfCov}: using unimodal alternative, 2000 DE genes, uninformative covariate, 10 vs. 10 comparison.
\item \code{yeastH-results-de10-uninfCov}: using unimodal alternative, 500 DE genes, uninformative covariate, 10 vs. 10 comparison.
\item \code{yeastII-results-de10-uninfCov}: using bimodal alternative, 2000 DE genes, uninformative covariate, 10 vs. 10 comparison.
\item \code{yeastIIH-results-de10-uninfCov}: using bimodal alternative, 500 DE genes, uninformative covariate, 10 vs. 10 comparison.
}
}
\examples{
\dontrun{
`yeast-results-de5`()
`yeast-results-de5-uninfCov`()
`yeast-results-de10`()
`yeast-results-de10-uninfCov`()
`yeast-results-null5`()
`yeast-results-null10`()
`yeastH-results-de5`()
`yeastH-results-de5-uninfCov`()
`yeastH-results-de10`()
`yeastH-results-de10-uninfCov`()
`yeastHW-results-de5`()
`yeastHW-results-de10`()
`yeastII-results-de5`()
`yeastII-results-de5-uninfCov`()
`yeastII-results-de10`()
`yeastII-results-de10-uninfCov`()
`yeastIIH-results-de5`()
`yeastIIH-results-de5-uninfCov`()
`yeastIIH-results-de10`()
`yeastIIH-results-de10-uninfCov`()
`yeastIIHW-results-de5`()
`yeastIIHW-results-de10`()
`yeastIIW-results-de5`()
`yeastIIW-results-de10`()
`yeastW-results-de5`()
`yeastW-results-de10`()
}

}
\references{
\itemize{
\item Schurch, N. J. et al. (2016). How many biological replicates are needed in an RNA-seq experiment and which differential expression tool should you use?. RNA, 22(6), 839-851.
}
}
\seealso{
\code{\link{polyester-insilico}}
}
